Reorganize source tree more:
* Remove mlcommons prefix
* Move ssd_resnet34, ssd_mobilenet, resnet50 to separate module from mlcommons
* Move common module from mlcommons to its parent
