+++ furiosa/models/vision/common/datasets/coco-label-paper.txt
