+++ furiosa/models/vision/common/datasets/coco-label-ssdmobilelarge.txt
