+++ furiosa/models/vision/common/datasets/coco.py
