+++ furiosa/models/vision/common/datasets/imagenet-1k-label.txt
