+++ furiosa/models/vision/common/datasets/imagenet1k.py
