+++ furiosa/models/vision/common/utils/encoder.py
