--- furiosa/models/vision/image_classification.py
@@ -1,1 +1,0 @@
-from .mlcommons.resnet50 import MLCommonsResNet50Model  # noqa: F401
