--- furiosa/models/vision/object_detection.py
@@ -1,3 +1,0 @@
-from .mlcommons.ssd_mobilenet_large import MLCommonsSSDLargeModel  # noqa: F401
-from .mlcommons.ssd_mobilenet_small import MLCommonsSSDSmallModel  # noqa: F401
-from .yolov5.large import YoloV5LargeModel  # noqa: F401
