+++ furiosa/models/vision/resnet50.py
