+++ scripts/test_resnet50.py
@@ -1,6 +1,6 @@
 import pytest
 
-from furiosa.models import vision
+from furiosa.models.vision import nonblocking
 from furiosa.registry import Model
 
 from .helpers.util import InferenceTestSessionWrapper
@@ -8,7 +8,7 @@ from .helpers.util import InferenceTestSessionWrapper
 
 @pytest.mark.asyncio
 async def test_mlcommons_resnet50_perf():
-    m: Model = await vision.MLCommonsResNet50()
+    m: Model = await nonblocking.ResNet50()
     test_image_path = "scripts/assets/cat.jpg"
 
     with InferenceTestSessionWrapper(m) as sess:
