+++ scripts/test_ssd_mobilenet_v1.py
@@ -1,30 +1,23 @@
 import numpy as np
 import pytest
 
-from furiosa.models import vision
+from furiosa.models.vision import nonblocking
 from furiosa.registry import Model
 
 from .helpers.util import InferenceTestSessionWrapper
 
 
 @pytest.mark.asyncio
-async def test_mlcommons_ssd_resnet34_perf():
-    m: Model = await vision.MLCommonsSSDResNet34()
+async def test_mlcommons_mobilessd_small_perf():
+    m: Model = await nonblocking.SSDMobileNet()
     test_image_path = "scripts/assets/cat.jpg"
 
-    assert len(m.classes) == 81, f"Classes is 81, but {len(m.classes)}"
+    assert len(m.classes) == 92, f"Classes is 92, but {len(m.classes)}"
     with InferenceTestSessionWrapper(m) as sess:
-        true_bbox = np.array(
-            [
-                [264.24792, 259.05603, 699.12964, 474.65332],
-                [221.0502, 123.12275, 549.879, 543.1015],
-            ],
-            dtype=np.float32,
-        )
-        true_classid = np.array([16, 16], dtype=np.int32)
-        true_confidence = np.array([0.37563688, 0.8747512], dtype=np.float32)
-
-        result = sess.inference(test_image_path, post_config={"confidence_threshold": 0.3})
+        true_bbox = np.array([[187.30786, 88.035324, 763.6886, 655.2937]], dtype=np.float32)
+        true_classid = np.array([17], dtype=np.int32)
+        true_confidence = np.array([0.97390455], dtype=np.float32)
+        result = sess.inference(test_image_path)
         assert len(result) == 3, "ssd_resnet34 output shape must be (1, 3)"
         bbox, classid, confidence = result
         assert np.array_equal(classid, true_classid), f"wrong classid: {classid}, expected 16(cat)"
