+++ tests/test_artifacts.py
@@ -3,7 +3,7 @@
 import pytest
 import yaml
 
-from furiosa.models import vision
+from furiosa.models.vision import nonblocking
 
 
 def sanity_check_for_dvc_file(model, dvc_file_path: str):
@@ -14,7 +14,7 @@ def sanity_check_for_dvc_file(model, dvc_file_path: str):
 @pytest.mark.asyncio
 async def test_mlcommons_resnet50():
     sanity_check_for_dvc_file(
-        await vision.MLCommonsResNet50(),
+        await nonblocking.ResNet50(),
         "models/mlcommons_resnet50_v1.5_int8.onnx.dvc",
     )
 
@@ -22,7 +22,7 @@ async def test_mlcommons_resnet50():
 @pytest.mark.asyncio
 async def test_mlcommons_ssd_mobilenet():
     sanity_check_for_dvc_file(
-        await vision.MLCommonsSSDMobileNet(),
+        await nonblocking.SSDMobileNet(),
         "models/mlcommons_ssd_mobilenet_v1_int8.onnx.dvc",
     )
 
@@ -30,7 +30,7 @@ async def test_mlcommons_ssd_mobilenet():
 @pytest.mark.asyncio
 async def test_mlcommons_ssd_resnet34():
     sanity_check_for_dvc_file(
-        await vision.MLCommonsSSDResNet34(),
+        await nonblocking.SSDResNet34(),
         "models/mlcommons_ssd_resnet34_int8.onnx.dvc",
     )
 
@@ -38,6 +38,6 @@ async def test_mlcommons_ssd_resnet34():
 @pytest.mark.asyncio
 async def test_yolov5_large():
     sanity_check_for_dvc_file(
-        await vision.YoloV5Large(),
+        await nonblocking.YOLOv5l(),
         "models/yolov5l_int8.onnx.dvc",
     )
